py/pyproject.toml
[tool.poetry]
name = "advent of code 2023 (python)"
version = "0.1.0"
description = "My solutions to AoC 2023 in Python"
authors = ["adrian-zuur <adrian.zuur@globant.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
